+++ ogcapi-processes/src/gdal_loader.rs
@@ -128,7 +128,7 @@ impl Processor for GdalLoader {
         "0.1.0"
     }
 
-    fn process(&self) -> Result<Process> {
+    async fn process(&self) -> Result<Process> {
         Process::try_new(
             self.id(),
             self.version(),
@@ -364,7 +364,7 @@ mod tests {
 
         println!(
             "Process:\n{}",
-            serde_json::to_string_pretty(&loader.process().unwrap()).unwrap()
+            serde_json::to_string_pretty(&loader.process().await.unwrap()).unwrap()
         );
 
         let input = GdalLoaderInputs {