+++ CHANGELOG.md
@@ -53,6 +53,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Set default item type of collection as `feature`.
 - Define numeric feature id as `u64`.
 - Remove default Crs implementation.
+- (`OGCAPI - Processes`): Made `process` functions in `Processor` trait async, so that implementations can perform asynchronous operations.
 
 ## [0.3.0] - 2025-04-05